Updated add_person function and organize function structures
