+++ RFID_Project/json practice/get_json_data.py
